Create README.md
Data used to see if there is a difference of USAPL competitors strength over a set period of time. Data is from USAPL rankings form 2014 to 2023, from OpenPowerlifting. Metrics used for this is the total overall weight from Squats, Deadlifts and Bench Press.

Data is organized for the top ten weightlifters by Total Weight of each exercise, and ranked highest to lowest. Removed any data that was not for the weight class of 198 and 205(no 198 weight class for 2015-2022, weight similar to 198 weight class) where the range of a competitors weight is from 181.81 lbs to 198.40 lbs.